moved Component to entity-manager.ts; starting findEntitySet
