+++ src/renderer.ts
@@ -1,4 +1,10 @@
-import { ComponentDef, EntityManager, EM, Entity } from "./entity-manager.js";
+import {
+  ComponentDef,
+  EntityManager,
+  EM,
+  Entity,
+  Component,
+} from "./entity-manager.js";
 import { ColorDef } from "./game/game.js";
 import {
   CameraDef,
@@ -12,8 +18,6 @@ import { Mesh, MeshHandle, MeshHandleDef } from "./mesh-pool.js";
 import { Motion, MotionDef } from "./phys_motion.js";
 import { tempQuat, tempVec } from "./temp-pool.js";
 import { TimeDef } from "./time.js";
-
-export type Component<DEF> = DEF extends ComponentDef<any, infer P> ? P : never;
 
 const SMOOTH = true;
